Refactor NewsInfo component to allow custom number of news to be shown
