+++ src/components/NewsInfo/NewsInfo.js
@@ -1,85 +1,22 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Link, Paper, Typography } from '@material-ui/core';
-import { FormattedMessage } from 'react-intl';
-import { getIcon } from '../SMIcon';
+import NewsItem from './components/NewsItem/NewsItem';
 
 const NewsInfo = ({
-  classes, getLocaleText, news,
+  showCount, news,
 }) => {
   if (!news.length) {
     return null;
   }
 
-  const {
-    title,
-    lead_paragraph: leadParagraph,
-    picture_url: pictureURL,
-    external_url_title: eUrlTitle,
-    external_url: eUrl,
-  } = news[0];
+  const dataToRender = news.slice(0, showCount);
 
-  const tTitle = title && getLocaleText(title);
-  const tLeadParagraph = leadParagraph && getLocaleText(leadParagraph);
-  const teUrlTitle = eUrlTitle && getLocaleText(eUrlTitle);
-  const urlHref = eUrl && eUrl !== '' ? getLocaleText(eUrl) : false;
-  const icon = getIcon('servicemapLogoIcon', {
-    className: classes.icon,
-  });
-  const imgSrc = pictureURL && pictureURL !== '' ? pictureURL : false;
-
-  return (
-    <Paper className={classes.container}>
-      <div className="row padding">
-        <Typography variant="srOnly" component="h3">
-          <FormattedMessage id="general.news.info.title" />
-        </Typography>
-        {icon}
-        <div className={`column ${classes.titleContainer}`}>
-          <Typography variant="subtitle1" component="h4">{tTitle}</Typography>
-        </div>
-      </div>
-      {
-        imgSrc
-        && (
-          <img alt="" src={imgSrc} className={classes.image} />
-        )
-      }
-
-      <div className={`column ${!imgSrc ? classes.hidePaddingTop : ''} ${classes.bottomContent}`}>
-        <Typography
-          align="left"
-          component="p"
-          variant="body2"
-        >
-          {tLeadParagraph}
-        </Typography>
-        {
-          teUrlTitle
-          && urlHref
-          && (
-            <Link href={urlHref} target="_blank">
-              <Typography align="left" variant="body2" component="p">{teUrlTitle}</Typography>
-            </Link>
-          )
-        }
-      </div>
-    </Paper>
-  );
+  return dataToRender.map(item => (
+    <NewsItem item={item} />
+  ));
 };
 
 NewsInfo.propTypes = {
-  classes: PropTypes.shape({
-    bottomContent: PropTypes.string,
-    container: PropTypes.string,
-    hidePaddingTop: PropTypes.string,
-    icon: PropTypes.string,
-    image: PropTypes.string,
-    title: PropTypes.string,
-    titleContainer: PropTypes.string,
-    subtitle: PropTypes.string,
-  }).isRequired,
-  getLocaleText: PropTypes.func.isRequired,
   news: PropTypes.arrayOf(PropTypes.shape({
     lead_paragraph: PropTypes.shape({
       fi: PropTypes.string,
@@ -95,6 +32,11 @@ NewsInfo.propTypes = {
       fi: PropTypes.string,
     }),
   })).isRequired,
+  showCount: PropTypes.number,
+};
+
+NewsInfo.defaultProps = {
+  showCount: 1,
 };
 
 export default NewsInfo;
